requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
torchinfo>=1.8.0
jupyter>=1.0.0
numpy>=1.24.0
wandb>=0.15.0
Pillow>=9.0.0
tqdm>=4.60.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
